pyproject.toml
[project]
name = "challenge-crud-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "faker>=40.11.1",
    "fastapi>=0.135.2",
    "fastapi-health>=0.4.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy-filters>=0.13.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
